+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/quill-renderer/compare/0.1.2...develop)
+## [Unreleased](https://github.com/ixocreate/quill-renderer/compare/0.1.3...develop)
+
+## [v0.1.3 (2019-10-22)](https://github.com/ixocreate/schema-package/compare/0.1.2...0.1.3)
+### Fixed
+- improved handling for malformed deltas without ending slash
+- fixed handling of linebreak=true deltas
 
 ## [v0.1.2 (2019-09-12)](https://github.com/ixocreate/schema-package/compare/0.1.1...0.1.2)
 ### Fixed